moved SB editor to own file and added DOM SB
